+++ packages/node-experimental/src/integrations/index.ts
@@ -26,6 +26,7 @@ export {
 
 export { Express } from './express';
 export { Http } from './http';
+export { NodeFetch } from './node-fetch';
 export { Fastify } from './fastify';
 export { GraphQL } from './graphql';
 export { Mongo } from './mongo';