+++ packages/node/src/anr/index.ts
@@ -1,43 +1,12 @@
 import type { Event, StackFrame } from '@sentry/types';
-import { logger } from '@sentry/utils';
+import { logger, watchdogTimer } from '@sentry/utils';
 import { spawn } from 'child_process';
-import * as inspector from 'inspector';
 
 import { addGlobalEventProcessor, captureEvent, flush } from '..';
 import { captureStackTrace } from './debugger';
 
 const DEFAULT_INTERVAL = 50;
 const DEFAULT_HANG_THRESHOLD = 5000;
-
-/**
- * A node.js watchdog timer
- * @param pollInterval The interval that we expect to get polled at
- * @param anrThreshold The threshold for when we consider ANR
- * @param callback The callback to call for ANR
- * @returns A function to call to reset the timer
- */
-function watchdogTimer(pollInterval: number, anrThreshold: number, callback: () => void): () => void {
-  let lastPoll = process.hrtime();
-  let triggered = false;
-
-  setInterval(() => {
-    const [seconds, nanoSeconds] = process.hrtime(lastPoll);
-    const diffMs = Math.floor(seconds * 1e3 + nanoSeconds / 1e6);
-
-    if (triggered === false && diffMs > pollInterval + anrThreshold) {
-      triggered = true;
-      callback();
-    }
-
-    if (diffMs < pollInterval + anrThreshold) {
-      triggered = false;
-    }
-  }, 20);
-
-  return () => {
-    lastPoll = process.hrtime();
-  };
-}
 
 interface Options {
   /**
@@ -67,7 +36,7 @@ interface Options {
    */
   captureStackTrace: boolean;
   /**
-   * Log debug information.
+   * @deprecated Use 'init' debug option instead
    */
   debug: boolean;
 }
@@ -98,12 +67,19 @@ function sendEvent(blockedMs: number, frames?: StackFrame[]): void {
   });
 }
 
+interface InspectorApi {
+  open: (port: number) => void;
+  url: () => string | undefined;
+}
+
 /**
  * Starts the node debugger and returns the inspector url.
  *
  * When inspector.url() returns undefined, it means the port is already in use so we try the next port.
  */
 function startInspector(startPort: number = 9229): string | undefined {
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  const inspector: InspectorApi = require('inspector');
   let inspectorUrl: string | undefined = undefined;
   let port = startPort;
 
@@ -118,9 +94,7 @@ function startInspector(startPort: number = 9229): string | undefined {
 
 function startChildProcess(options: Options): void {
   function log(message: string, ...args: unknown[]): void {
-    if (options.debug) {
-      logger.log(`[ANR] ${message}`, ...args);
-    }
+    logger.log(`[ANR] ${message}`, ...args);
   }
 
   try {
@@ -135,7 +109,7 @@ function startChildProcess(options: Options): void {
 
     const child = spawn(process.execPath, [options.entryScript], {
       env,
-      stdio: options.debug ? ['inherit', 'inherit', 'inherit', 'ipc'] : ['ignore', 'ignore', 'ignore', 'ipc'],
+      stdio: logger.isEnabled() ? ['inherit', 'inherit', 'inherit', 'ipc'] : ['ignore', 'ignore', 'ignore', 'ipc'],
     });
     // The child process should not keep the main process alive
     child.unref();
@@ -166,9 +140,7 @@ function startChildProcess(options: Options): void {
 
 function handleChildProcess(options: Options): void {
   function log(message: string): void {
-    if (options.debug) {
-      logger.log(`[ANR child process] ${message}`);
-    }
+    logger.log(`[ANR child process] ${message}`);
   }
 
   process.title = 'sentry-anr';
@@ -210,10 +182,10 @@ function handleChildProcess(options: Options): void {
     }
   }
 
-  const ping = watchdogTimer(options.pollInterval, options.anrThreshold, watchdogTimeout);
+  const { poll } = watchdogTimer(options.pollInterval, options.anrThreshold, watchdogTimeout);
 
   process.on('message', () => {
-    ping();
+    poll();
   });
 }
 
@@ -257,6 +229,7 @@ export function enableAnrDetection(options: Partial<Options>): Promise<void> {
     pollInterval: options.pollInterval || DEFAULT_INTERVAL,
     anrThreshold: options.anrThreshold || DEFAULT_HANG_THRESHOLD,
     captureStackTrace: !!options.captureStackTrace,
+    // eslint-disable-next-line deprecation/deprecation
     debug: !!options.debug,
   };