+++ packages/react/src/sdk.ts
@@ -1,12 +1,17 @@
 import type { BrowserOptions } from '@sentry/browser';
 import { init as browserInit, SDK_VERSION } from '@sentry/browser';
+import type { SdkMetadata } from '@sentry/types';
 
 /**
  * Inits the React SDK
  */
 export function init(options: BrowserOptions): void {
-  options._metadata = options._metadata || {};
-  options._metadata.sdk = options._metadata.sdk || {
+  const opts = {
+    _metadata: {} as SdkMetadata,
+    ...options,
+  };
+
+  opts._metadata.sdk = opts._metadata.sdk || {
     name: 'sentry.javascript.react',
     packages: [
       {
@@ -16,5 +21,5 @@ export function init(options: BrowserOptions): void {
     ],
     version: SDK_VERSION,
   };
-  browserInit(options);
+  browserInit(opts);
 }