+++ packages/sveltekit/src/server/index.ts
@@ -6,6 +6,7 @@
 export {
   addGlobalEventProcessor,
   addBreadcrumb,
+  addIntegration,
   captureException,
   captureEvent,
   captureMessage,
@@ -51,6 +52,7 @@ export {
   startActiveSpan,
   startInactiveSpan,
   startSpanManual,
+  continueTrace,
 } from '@sentry/node';
 
 // We can still leave this for the carrier init and type exports